Don't try to compile cargo-credential-gnome-secret on non-Linux platforms.
