+++ credential/cargo-credential-gnome-secret/build.rs
@@ -1,3 +1,8 @@
 fn main() {
-    pkg_config::probe_library("libsecret-1").unwrap();
+    if cfg!(target_os = "linux") {
+        // TODO: Consider ignoring errors when libsecret is not installed and
+        // switching the impl to UnsupportedCredential (possibly along with a
+        // warning?).
+        pkg_config::probe_library("libsecret-1").unwrap();
+    }
 }
